+++ judo-runtime-core-dao-rdbms/src/main/java/hu/blackbelt/judo/runtime/core/dao/rdbms/query/RdbmsBuilder.java
@@ -25,27 +25,21 @@ import hu.blackbelt.judo.dispatcher.api.VariableResolver;
 import hu.blackbelt.judo.meta.asm.runtime.AsmUtils;
 import hu.blackbelt.judo.meta.query.*;
 import hu.blackbelt.judo.meta.rdbms.runtime.RdbmsModel;
-import hu.blackbelt.judo.meta.rdbmsRules.Rule;
 import hu.blackbelt.judo.meta.rdbmsRules.Rules;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.*;
-import hu.blackbelt.judo.runtime.core.dao.rdbms.executors.SelectStatementExecutor;
-import hu.blackbelt.judo.runtime.core.dao.rdbms.executors.StatementExecutor;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.mappers.MapperFactory;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.mappers.RdbmsMapper;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.*;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.join.*;
-import hu.blackbelt.judo.runtime.core.dao.rdbms.query.utils.RdbmsAliasUtil;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.query.processor.*;
 import hu.blackbelt.mapper.api.Coercer;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.emf.common.util.*;
 import org.eclipse.emf.ecore.*;
 
-import java.text.CharacterIterator;
-import java.text.StringCharacterIterator;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static com.google.common.base.Preconditions.checkArgument;
@@ -90,6 +84,15 @@ public class RdbmsBuilder<ID> {
     @Getter
     private final Dialect dialect;
 
+
+    private final AncestorJoinsProcessor ancestorJoinsProcessor;
+    private final CastJoinProcessor castJoinProcessor;
+    private final ContainerJoinProcessor containerJoinProcessor;
+    private final CustomJoinProcessor customJoinProcessor;
+    private final FilterJoinProcessor filterJoinProcessor;
+    private final SimpleJoinProcessor simpleJoinProcessor;
+    private final SubSelectJoinProcessor subSelectJoinProcessor;
+
     private final ThreadLocal<Map<String, Collection<? extends RdbmsField>>> CONSTANT_FIELDS = new ThreadLocal<>();
 
     @Builder
@@ -122,6 +125,39 @@ public class RdbmsBuilder<ID> {
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("Rules not found in RDBMS model"));
 
+        this.simpleJoinProcessor = SimpleJoinProcessor.builder()
+                .rdbmsResolver(rdbmsResolver)
+                .rules(rules)
+                .ancestorNameFactory(ancestorNameFactory)
+                .build();
+
+        this.ancestorJoinsProcessor = AncestorJoinsProcessor.builder()
+                .ancestorNameFactory(ancestorNameFactory)
+                .descendantNameFactory(descendantNameFactory)
+                .rdbmsResolver(rdbmsResolver)
+                .build();
+
+        this.castJoinProcessor = CastJoinProcessor.builder()
+                .rdbmsResolver(rdbmsResolver)
+                .build();
+
+        this.containerJoinProcessor = ContainerJoinProcessor.builder()
+                .rdbmsResolver(rdbmsResolver)
+                .build();
+
+        this.customJoinProcessor = CustomJoinProcessor.builder()
+                .rdbmsResolver(rdbmsResolver)
+                .asmUtils(asmUtils)
+                .build();
+
+        this.filterJoinProcessor = FilterJoinProcessor.builder()
+                .rdbmsResolver(rdbmsResolver)
+                .build();
+
+
+        this.subSelectJoinProcessor = SubSelectJoinProcessor.builder()
+                .rdbmsResolver(rdbmsResolver)
+                .build();
     }
 
     @SuppressWarnings("unchecked")
@@ -148,30 +184,6 @@ public class RdbmsBuilder<ID> {
     }
 
 
-    @Builder(toBuilder = true)
-    public static class RdbmsBuilderContext {
-        @NonNull
-        public RdbmsBuilder<?> rdbmsBuilder;
-        @NonNull
-        public EMap<Node, EList<EClass>> ancestors;
-        @NonNull
-        public EMap<Node, EList<EClass>> descendants;
-        public SubSelect parentIdFilterQuery;
-        public Map<String, Object> queryParameters;
-    }
-
-    @Builder
-    public static class ProcessJoinParameters {
-        @NonNull
-        Node join;
-
-        boolean withoutFeatures;
-        Map<String, Object> mask;
-
-        @NonNull
-        RdbmsBuilderContext builderContext;
-    }
-
     /**
      * Resolve logical JOIN and return RDBMS JOIN definition.
      *
@@ -179,11 +191,12 @@ public class RdbmsBuilder<ID> {
      * @return RDBMS JOIN definition(s)
      */
     @SuppressWarnings("unchecked")
-    public List<RdbmsJoin> processJoin(ProcessJoinParameters params) {
-        final Node join = params.join;
-        final boolean withoutFeatures = params.withoutFeatures;
-        final Map<String, Object> mask = params.mask;
-        final RdbmsBuilderContext builderContext = params.builderContext;
+    public List<RdbmsJoin> processJoin(JoinProcessParameters params) {
+        final Node join = params.getJoin();
+        final boolean withoutFeatures = params.isWithoutFeatures();
+        final Map<String, Object> mask = params.getMask();
+        final RdbmsBuilderContext builderContext = params.getBuilderContext();
+
         final RdbmsBuilder<?> rdbmsBuilder = builderContext.rdbmsBuilder;
         final EMap<Node, EList<EClass>> ancestors = builderContext.ancestors;
         final SubSelect parentIdFilterQuery = builderContext.parentIdFilterQuery;
@@ -191,506 +204,47 @@ public class RdbmsBuilder<ID> {
 
 
         if (join instanceof ReferencedJoin) {
-            return processSimpleJoin(ProcessSimpleJoinParameters.builder()
+            return simpleJoinProcessor.process(SimpleJoinProcessorParameters.builder()
                     .join((ReferencedJoin) join)
                     .reference(((ReferencedJoin)  join).getReference())
                     .opposite(((ReferencedJoin)  join).getReference().getEOpposite())
                     .builderContext(builderContext)
                     .build());
         } else if (join instanceof ContainerJoin) {
-            return processContainerJoin((ContainerJoin) join, builderContext);
+            return containerJoinProcessor.process(
+                    (ContainerJoin) join,
+                    builderContext);
         } else if (join instanceof CastJoin) {
-            return processCastJoin(join, builderContext);
+            return castJoinProcessor.process(join,
+                    builderContext);
         } else if (join instanceof SubSelectJoin) {
-            final SubSelect subSelect = ((SubSelectJoin) join).getSubSelect();
-            subSelect.getFilters().addAll(join.getFilters());
-
-            final List<RdbmsJoin> joins = new ArrayList<>();
-            final Map<String, Object> _mask = mask != null && subSelect.getTransferRelation() != null ? (Map<String, Object>) mask.get(subSelect.getTransferRelation().getName()) : null;
-            final RdbmsResultSet<ID> resultSetHandler =
-                    RdbmsResultSet.<ID>builder()
-                            .query(subSelect)
-                            .filterByInstances(false)
-                            .parentIdFilterQuery(parentIdFilterQuery)
-                            .rdbmsBuilder((RdbmsBuilder<ID>) rdbmsBuilder)
-                            .seek(null)
-                            .withoutFeatures(withoutFeatures)
-                            .mask(_mask)
-                            .queryParameters(queryParameters)
-                            .skipParents(false)
-                            .build();
-
-            if (!AsmUtils.equals(((SubSelectJoin) join).getPartner(), subSelect.getBase())) {
-                joins.add(RdbmsTableJoin.builder()
-                        .tableName(rdbmsBuilder.getTableName(subSelect.getBase().getType()))
-                        .columnName(StatementExecutor.ID_COLUMN_NAME)
-                        .partnerTable(((SubSelectJoin) join).getPartner())
-                        .partnerColumnName(StatementExecutor.ID_COLUMN_NAME)
-                        .alias(subSelect.getBase().getAlias())
-                        .build());
-            }
-
-            joins.add(RdbmsQueryJoin.<ID>builder()
-                    .resultSet(resultSetHandler)
-                    .outer(true)
-                    .columnName(RdbmsAliasUtil.getOptionalParentIdColumnAlias(subSelect.getContainer()))
-//                    .partnerTable(!subSelect.getNavigationJoins().isEmpty() && AsmUtils.equals(subSelect.getNavigationJoins().get(0).getPartner(), join) ? subSelect.getBase() : null)
-//                    .partnerColumnName(!subSelect.getNavigationJoins().isEmpty() && AsmUtils.equals(subSelect.getNavigationJoins().get(0).getPartner(), join) ? StatementExecutor.ID_COLUMN_NAME : null)
-                    .partnerTable(subSelect.getBase())
-                    .partnerColumnName(StatementExecutor.ID_COLUMN_NAME)
-                    .alias(subSelect.getAlias())
-                    .build());
-            final Optional<Feature> selectorFeature = subSelect.getSelect().getFeatures().stream()
-                    .filter(f -> (f instanceof Function) && (FunctionSignature.MIN_INTEGER.equals(((Function) f).getSignature()) || FunctionSignature.MAX_INTEGER.equals(((Function) f).getSignature())))
-                    .findAny();
-            checkArgument(selectorFeature.isPresent(), "SubSelectFeature of head/tail/any must exists");
-            final Optional<RdbmsMapper.RdbmsTarget> selectorTarget = RdbmsMapper.getTargets(selectorFeature.get()).findAny();
-            checkArgument(selectorTarget.isPresent(), "SubSelectFeature target must exists");
-            joins.add(RdbmsTableJoin.builder()
-                    .tableName(rdbmsResolver.rdbmsTable(join.getType()).getSqlName())
-                    .alias(join.getAlias())
-                    .columnName(StatementExecutor.ID_COLUMN_NAME)
-                    .partnerTable(subSelect)
-                    .partnerColumnName(selectorTarget.get().getAlias() + "_" + selectorTarget.get().getTarget().getIndex())
-                    .outer(true)
-                    .build());
-
-            if (ancestors.containsKey(join)) {
-                addAncestorJoins(joins, join, ancestors);
-            }
-
-//            if (descendants.containsKey(join)) {
-//                result.addAll(getDescendantJoins(join, descendants, result));
-//
-////                result.addAll(descendants.get(join).stream()
-////                        .flatMap(descendant -> getDescendantJoins(join, descendants, result).stream())
-////                        .collect(Collectors.toList()));
-//            }
-
-            return joins;
+            return subSelectJoinProcessor.process(
+                    (SubSelectJoin) join,
+                    withoutFeatures,
+                    mask,
+                    builderContext);
         } else if (join instanceof CustomJoin) {
-            final List<RdbmsJoin> joins = new ArrayList<>();
-
             final CustomJoin customJoin = (CustomJoin) join;
-
-            final String sql;
-            if (customJoin.getNavigationSql().indexOf('`') != -1) {
-                sql = resolveRdbmsNames(customJoin.getNavigationSql());
-            } else {
-                sql = customJoin.getNavigationSql();
-            }
-
-            joins.add(RdbmsCustomJoin.builder()
-                    .sql(sql)
-                    .sourceIdSetParameterName(customJoin.getSourceIdSetParameter())
-                    .alias(join.getAlias())
-                    .columnName(customJoin.getSourceIdParameter())
-                    .partnerTable(customJoin.getPartner())
-                    .partnerColumnName(StatementExecutor.ID_COLUMN_NAME)
-                    .outer(true)
-                    .build());
-
-            if (ancestors.containsKey(join)) {
-                addAncestorJoins(joins, join, ancestors);
-            }
-            return joins;
+            return customJoinProcessor.process(
+                    customJoin,
+                    builderContext);
         } else {
             throw new IllegalStateException("Invalid JOIN");
         }
     }
 
-    private String resolveRdbmsNames(final String sql) {
-        final StringBuilder result = new StringBuilder();
-        final StringCharacterIterator it = new StringCharacterIterator(sql);
-
-        boolean resolving = false;
-        StringBuilder fqNameBuilder = null;
-        for (char ch = it.first(); ch != CharacterIterator.DONE; ch = it.next()) {
-            if (ch == '`' && resolving) {
-                final String fqName = fqNameBuilder.toString();
-
-                final Optional<EAttribute> attribute = asmUtils.resolveAttribute(fqName);
-                final Optional<EReference> reference = attribute.isPresent() ? Optional.empty() : asmUtils.resolveReference(fqName);
-                final Optional<EClass> type = attribute.isPresent() || reference.isPresent() ? Optional.empty() : asmUtils.resolve(fqName).filter(c -> c instanceof EClass).map(c -> (EClass) c);
-
-                // TODO - support resolving junction table names
-                if (attribute.isPresent()) {
-                    result.append(rdbmsResolver.rdbmsField(attribute.get()).getSqlName());
-                } else if (reference.isPresent()) {
-                    result.append(rdbmsResolver.rdbmsField(reference.get()).getSqlName());
-                } else if (type.isPresent()) {
-                    result.append(rdbmsResolver.rdbmsTable(type.get()).getSqlName());
-                } else {
-                    throw new IllegalStateException("Unable to resolve ASM element name: " + fqName);
-                }
-
-                resolving = false;
-            } else if (ch == '`' && !resolving) {
-                fqNameBuilder = new StringBuilder();
-                resolving = true;
-            } else if (resolving) {
-                fqNameBuilder.append(ch);
-            } else {
-                result.append(ch);
-            }
-        }
-
-        if (resolving) {
-            log.error("SQL syntax is invalid (terminated while resolving RDBMS name): {}", sql);
-            throw new IllegalArgumentException("Invalid custom SQL");
-        }
-
-        return result.toString();
-    }
-
-    @SuppressWarnings("unchecked")
-    @Builder
-    private static class ProcessSimpleJoinParameters {
-        @Builder.Default
-        String postfix = "";
-        @NonNull
-        Join join;
-
-        EReference reference;
-        EReference opposite;
-
-        @NonNull
-        RdbmsBuilderContext builderContext;
-    }
-
-//final String postfix, final Join join, final EReference reference, final EReference opposite, final EMap<Node, EList<EClass>> ancestors, final EMap<Node, EList<EClass>> descendants, final SubSelect parentIdFilterQuery, final Map<String, Object> queryParameters
-    private List<RdbmsJoin> processSimpleJoin(ProcessSimpleJoinParameters params) {
-        final String postfix = params.postfix;
-        final Join join = params.join;
-        final EReference reference = params.reference;
-        final EReference opposite = params.opposite;
-        final RdbmsBuilderContext builderContext = params.builderContext;
-        final EMap<Node, EList<EClass>> ancestors = builderContext.ancestors;
-
-
-        final EClass targetType = join.getType();
-        final Node node = join.getPartner();
-        final EClass sourceType = node.getType();
-
-        if (log.isTraceEnabled()) {
-            log.trace(" => processing JOIN: {}", join);
-            log.trace("    target type: {}", targetType.getName());
-            log.trace("    source type: {}", sourceType.getName());
-        }
-
-        final String tableName = rdbmsResolver.rdbmsTable(targetType).getSqlName();
-
-        // create builder for RDBMS JOIN definition
-        @SuppressWarnings("rawtypes")
-        final RdbmsJoin.RdbmsJoinBuilder builder = RdbmsTableJoin.builder()
-                .outer(true)
-                .tableName(tableName)
-                .alias(join.getAlias() + postfix)
-                .partnerTable(node);
-
-        final Rule rule;
-        EClass sourceContainer = null;
-        if (reference != null) {
-            // get RDBMS rule of a given reference
-            rule = rules.getRuleFromReference(reference);
-
-            sourceContainer = reference.getEContainingClass();
-            if (log.isTraceEnabled()) {
-                log.trace("    reference: {}", reference.getName());
-                log.trace("    reference container: {}", sourceContainer.getName());
-            }
-        } else {
-            rule = null;
-        }
-
-        final Rule oppositeRule;
-        if (opposite != null) {
-            oppositeRule = rules.getRuleFromReference(opposite);
-
-            final EClass oppositeContainer = opposite.getEReferenceType();
-            if (log.isTraceEnabled()) {
-                log.trace("    opposite: {}", opposite.getName());
-                log.trace("    opposite reference container: {}", oppositeContainer.getName());
-            }
-
-            if (sourceContainer == null) {
-                sourceContainer = oppositeContainer;
-            }
-        } else {
-            oppositeRule = null;
-        }
-
-        if (!AsmUtils.equals(sourceType, sourceContainer)) { // reference is inherited from another class, resolve ancestor too
-            log.trace("  - reference '{}' is inherited");
-
-            if (!ancestors.containsKey(node)) {
-                ancestors.put(node, new UniqueEList<>());
-            }
-            ancestors.get(node).add(sourceContainer);
-            builder.partnerTablePostfix(getAncestorPostfix(sourceContainer));
-        }
-        if (!targetType.getEAllSuperTypes().isEmpty()) {
-            if (!ancestors.containsKey(join)) {
-                ancestors.put(join, new UniqueEList<>());
-            }
-            ancestors.get(join).addAll(targetType.getEAllSuperTypes());
-        }
-
-        if (rule != null && rule.isForeignKey()) { // reference is owned by source class, target class has reference to the ID with different name
-            log.trace("  - reference '{}' is foreign key", reference.getName());
-
-            builder.columnName(StatementExecutor.ID_COLUMN_NAME).partnerColumnName(rdbmsResolver.rdbmsField(reference).getSqlName());
-        } else if (rule != null && rule.isInverseForeignKey()) {  // reference is owned by target class, source class has reference to the ID with different name
-            log.trace("  - reference '{}' is inverse foreign key", reference.getName());
-
-            builder.columnName(rdbmsResolver.rdbmsField(reference).getSqlName()).partnerColumnName(StatementExecutor.ID_COLUMN_NAME);
-        } else if (rule != null && rule.isJoinTable()) { // JOIN tables are not supported yet
-            log.trace("  - reference '{}' is JOIN table", reference.getName());
-
-            builder.columnName(StatementExecutor.ID_COLUMN_NAME).partnerColumnName(StatementExecutor.ID_COLUMN_NAME)
-                    .junctionTableName(rdbmsResolver.rdbmsJunctionTable(reference).getSqlName())
-                    .junctionColumnName(rdbmsResolver.rdbmsJunctionField(reference).getSqlName())
-                    .junctionOppositeColumnName(rdbmsResolver.rdbmsJunctionOppositeField(reference).getSqlName());
-        } else if (oppositeRule != null && oppositeRule.isForeignKey()) { // reference is owned by source class, target class has reference to the ID with different name (defined by opposite reference)
-            log.trace("  - opposite reference '{}' is foreign key", opposite.getName());
-
-            builder.columnName(rdbmsResolver.rdbmsField(opposite).getSqlName()).partnerColumnName(StatementExecutor.ID_COLUMN_NAME);
-        } else if (oppositeRule != null && oppositeRule.isInverseForeignKey()) {  // reference is owned by target class, source class has reference to the ID with different name (defined by opposite reference)
-            log.trace("  - opposite reference '{}' is inverse foreign key", opposite.getName());
-
-            builder.columnName(StatementExecutor.ID_COLUMN_NAME).partnerColumnName(rdbmsResolver.rdbmsField(opposite).getSqlName());
-        } else if (oppositeRule != null && oppositeRule.isJoinTable()) { // JOIN tables are not supported yet
-            log.trace("  - opposite reference '{}' is JOIN table", opposite.getName());
-
-            builder.columnName(StatementExecutor.ID_COLUMN_NAME).partnerColumnName(StatementExecutor.ID_COLUMN_NAME)
-                    .junctionTableName(rdbmsResolver.rdbmsJunctionTable(opposite).getSqlName())
-                    .junctionColumnName(rdbmsResolver.rdbmsJunctionField(opposite).getSqlName())
-                    .junctionOppositeColumnName(rdbmsResolver.rdbmsJunctionOppositeField(opposite).getSqlName());
-        } else {
-            throw new IllegalStateException("Invalid reference");
-        }
-
-        if (!join.getFilters().isEmpty() && join.getFilters().stream().noneMatch(filter -> filter.getFeatures().stream().anyMatch(feature -> feature instanceof SubSelectFeature))) {
-            builder.onConditions(join.getFilters().stream()
-                    .map(f -> RdbmsFunction.builder()
-                            .pattern("EXISTS ({0})")
-                            .parameter(
-                                    RdbmsNavigationFilter.<ID>builder()
-                                            .builderContext(builderContext)
-                                            .filter(f)
-                                            .build())
-                            .build())
-                    .collect(Collectors.toList()));
-        }
-
-        final RdbmsJoin rdbmsJoin = builder.build();
-
-        final List<RdbmsJoin> joins = new ArrayList<>();
-        joins.add(rdbmsJoin);
-
-        if (ancestors.containsKey(join)) {
-            addAncestorJoins(joins, join, ancestors);
-        }
-
-        return joins;
-    }
-
-// final ContainerJoin join, final EMap<Node, EList<EClass>> ancestors,
-//                                                 final EMap<Node, EList<EClass>> descendants, final SubSelect parentIdFilterQuery,
-//                                                 final Map<String, Object> queryParameters
-
-    private List<RdbmsJoin> processContainerJoin(ContainerJoin join, RdbmsBuilderContext builderContext) {
-        final RdbmsBuilder<?> rdbmsBuilder = builderContext.rdbmsBuilder;
-        final EMap<Node, EList<EClass>> ancestors = builderContext.ancestors;
-        final EMap<Node, EList<EClass>> descendants = builderContext.descendants;
-        final SubSelect parentIdFilterQuery = builderContext.parentIdFilterQuery;
-        final Map<String, Object> queryParameters = builderContext.queryParameters;
-
-        final EClass targetType = join.getType();
-        final Node node = join.getPartner();
-        final EList<EReference> references = join.getReferences();
-        final EClass sourceType = node != null ? node.getType() : references.get(0).getEReferenceType();
-
-        if (log.isTraceEnabled()) {
-            log.trace(" => processing JOIN: {}", join);
-            log.trace("    target type: {}", targetType.getName());
-            log.trace("    source type: {}", sourceType.getName());
-            log.trace("    references: {}", references.stream().map(r -> AsmUtils.getReferenceFQName(r)).collect(Collectors.joining(", ")));
-        }
-
-        final List<RdbmsJoin> result = new ArrayList<>();
-        int index = 0;
-        for (final EReference r : join.getReferences()) {
-            result.addAll(processSimpleJoin(ProcessSimpleJoinParameters.builder()
-                    .postfix(RdbmsContainerJoin.POSTFIX + index++)
-                    .join(join)
-                    .opposite(r)
-                    .builderContext(builderContext)
-                    .build()));
-        }
-
-        result.add(RdbmsContainerJoin.builder()
-                .outer(true)
-                .tableName(rdbmsResolver.rdbmsTable(targetType).getSqlName())
-                .alias(join.getAlias())
-                .partnerTable(node)
-                .columnName(StatementExecutor.ID_COLUMN_NAME)
-                .references(references)
-                .partnerColumnName(StatementExecutor.ID_COLUMN_NAME)
-                .build());
-
-        return result;
-    }
-
-    private List<RdbmsJoin> processCastJoin(Node join, RdbmsBuilderContext builderContext) {
-        final RdbmsBuilder<?> rdbmsBuilder = builderContext.rdbmsBuilder;
-
-        EClass castTargetType = join.getType();
-        Set<EClass> typeSet = new HashSet<>(castTargetType.getEAllSuperTypes());
-        typeSet.add(castTargetType);
-
-        List<EClass> types = typeSet.stream()
-                                    .sorted((l, r) -> {
-                                        // ascending order
-                                        // type < supertype
-                                        if (l.getEAllSuperTypes().contains(r)) {
-                                            return -1;
-                                        }
-                                        if (r.getEAllSuperTypes().contains(l)) {
-                                            return 1;
-                                        }
-                                        return 0;
-                                    })
-                                    .collect(Collectors.toUnmodifiableList());
-
-        List<RdbmsJoin> rdbmsTableJoins = new ArrayList<>();
-        for (EClass type : types) {
-            String alias = join.getAlias();
-            Node partnerTable = ((CastJoin) join).getPartner();
-            List<RdbmsFunction> onConditions = new ArrayList<>();
-            RdbmsTableJoin rdbmsPartnerTable = null;
-
-            if (rdbmsTableJoins.isEmpty()) {
-                // original, first join
-                List<Filter> joinFilters = join.getFilters();
-                boolean joinFiltersWithoutSubSelectFeatures =
-                        !joinFilters.isEmpty()
-                        && joinFilters.stream()
-                                .noneMatch(filter -> filter.getFeatures().stream()
-                                                                 .anyMatch(feature -> feature instanceof SubSelectFeature));
-                if (joinFiltersWithoutSubSelectFeatures) {
-                    onConditions = joinFilters.stream()
-                            .map(f -> RdbmsFunction.builder()
-                                    .pattern("EXISTS ({0})")
-                                    .parameter(RdbmsNavigationFilter.<ID>builder()
-                                            .builderContext(builderContext)
-                                            .filter(f)
-                                            .build())
-                                    .build())
-                            .collect(Collectors.toList());
-                }
-            } else {
-                // additional join for supertypes
-                alias += rdbmsBuilder.getAncestorPostfix(type);
-                partnerTable = null;
-                rdbmsPartnerTable = (RdbmsTableJoin) rdbmsTableJoins.get(rdbmsTableJoins.size() - 1);
-            }
-
-            rdbmsTableJoins.add(RdbmsTableJoin.builder()
-                                              .outer(true)
-                                              .alias(alias)
-                                              .partnerTable(partnerTable)
-                                              .onConditions(onConditions)
-                                              .rdbmsPartnerTable(rdbmsPartnerTable)
-                                              .columnName(SelectStatementExecutor.ID_COLUMN_NAME)
-                                              .tableName(rdbmsResolver.rdbmsTable(type).getSqlName())
-                                              .partnerColumnName(SelectStatementExecutor.ID_COLUMN_NAME)
-                                              .build());
-        }
-
-        return rdbmsTableJoins;
-    }
-
 
     public void addAncestorJoins(final Collection<RdbmsJoin> joins, final Node node, final EMap<Node, EList<EClass>> ancestors) {
-        final EList<EClass> list;
-        if (ancestors.containsKey(node)) {
-            list = ancestors.get(node);
-        } else if (node.eContainer() instanceof SubSelect && ancestors.containsKey(((SubSelect) node.eContainer()).getSelect())) {
-            list = ancestors.get(((SubSelect) node.eContainer()).getSelect());
-        } else if (node.eContainer() instanceof Node && ancestors.containsKey(node.eContainer())) {
-            list = ancestors.get(node.eContainer());
-        } else {
-            list = ECollections.emptyEList();
-        }
-
-        List<RdbmsJoin> newJoins = list.stream()
-                .filter(c -> joins.stream().noneMatch(j -> Objects.equals(node.getAlias() + getAncestorPostfix(c), j.getAlias())))
-                .map(ancestor -> RdbmsTableJoin.builder()
-                        .tableName(rdbmsResolver.rdbmsTable(ancestor).getSqlName())
-                        .alias(node.getAlias() + getAncestorPostfix(ancestor))
-                        .columnName(StatementExecutor.ID_COLUMN_NAME)
-                        .partnerTable(node)
-                        .partnerColumnName(StatementExecutor.ID_COLUMN_NAME)
-                        .outer(true)
-                        .build())
-                .collect(Collectors.toList());
-
-        joins.addAll(newJoins);
+        ancestorJoinsProcessor.addAncestorJoins(joins, node, ancestors);
     }
 
-    public void addDescendantJoins(final Collection<RdbmsJoin> joins, final Node node, final EMap<Node, EList<EClass>> descendants) {
-        final EList<EClass> list;
-        if (descendants.containsKey(node)) {
-            list = descendants.get(node);
-        } else if (node.eContainer() instanceof SubSelect && descendants.containsKey(((SubSelect) node.eContainer()).getSelect())) {
-            list = descendants.get(((SubSelect) node.eContainer()).getSelect());
-        } else if (node.eContainer() instanceof Node && descendants.containsKey(node.eContainer())) {
-            list = descendants.get(node.eContainer());
-        } else {
-            list = ECollections.emptyEList();
-        }
-
-        List<RdbmsJoin> newJoins = list.stream()
-                .filter(c -> joins.stream().noneMatch(j -> Objects.equals(node.getAlias() + getDescendantPostfix(c), j.getAlias())))
-                .map(descendant -> RdbmsTableJoin.builder()
-                        .tableName(rdbmsResolver.rdbmsTable(descendant).getSqlName())
-                        .alias(node.getAlias() + getDescendantPostfix(descendant))
-                        .columnName(StatementExecutor.ID_COLUMN_NAME)
-                        .partnerTable(node)
-                        .partnerColumnName(StatementExecutor.ID_COLUMN_NAME)
-                        .outer(true)
-                        .build())
-                .collect(Collectors.toList());
-
-        joins.addAll(newJoins);
+    public void addFilterJoinsAndConditions(FilterJoinProcessorParameters params, RdbmsBuilderContext builderContext) {
+        filterJoinProcessor.process(params, builderContext);
     }
 
-    public Collection<RdbmsJoin> getDescendantJoins(final Node node, final EMap<Node, EList<EClass>> descendants, final Collection<RdbmsJoin> joins) {
-        final EList<EClass> list;
-        if (descendants.containsKey(node)) {
-            list = descendants.get(node);
-        } else if (node.eContainer() instanceof SubSelect && descendants.containsKey(((SubSelect) node.eContainer()).getSelect())) {
-            list = descendants.get(((SubSelect) node.eContainer()).getSelect());
-        } else if (node.eContainer() instanceof Node && descendants.containsKey(node.eContainer())) {
-            list = descendants.get(node.eContainer());
-        } else {
-            list = ECollections.emptyEList();
-        }
-
-        return list.stream()
-                .filter(c -> joins.stream().noneMatch(j -> Objects.equals(node.getAlias() + getDescendantPostfix(c), j.getAlias())))
-                .map(descendant -> RdbmsTableJoin.builder()
-                        .tableName(rdbmsResolver.rdbmsTable(descendant).getSqlName())
-                        .alias(node.getAlias() + getDescendantPostfix(descendant))
-                        .columnName(StatementExecutor.ID_COLUMN_NAME)
-                        .partnerTable(node)
-                        .partnerColumnName(StatementExecutor.ID_COLUMN_NAME)
-                        .outer(true)
-                        .build())
-                .collect(Collectors.toList());
+    public List<RdbmsJoin> processSimpleJoin(SimpleJoinProcessorParameters params) {
+        return simpleJoinProcessor.process(params);
     }
-
     public ThreadLocal<Map<String, Collection<? extends RdbmsField>>> getConstantFields() {
         return CONSTANT_FIELDS;
     }
