+++ judo-runtime-core-dao-rdbms/src/main/java/hu/blackbelt/judo/runtime/core/dao/rdbms/query/mappers/TypeAttributeMapper.java
@@ -22,10 +22,9 @@ package hu.blackbelt.judo.runtime.core.dao.rdbms.query.mappers;
 
 import hu.blackbelt.judo.meta.asm.runtime.AsmUtils;
 import hu.blackbelt.judo.meta.query.Node;
-import hu.blackbelt.judo.meta.query.SubSelect;
 import hu.blackbelt.judo.meta.query.TypeAttribute;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.executors.StatementExecutor;
-import hu.blackbelt.judo.runtime.core.dao.rdbms.query.RdbmsBuilder;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.query.RdbmsBuilderContext;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.RdbmsColumn;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.emf.common.util.EList;
@@ -34,25 +33,24 @@ import org.eclipse.emf.common.util.UniqueEList;
 import org.eclipse.emf.ecore.EClass;
 
 import java.util.Arrays;
-import java.util.Map;
 import java.util.stream.Stream;
 
 @Slf4j
 public class TypeAttributeMapper extends RdbmsMapper<TypeAttribute> {
 
     @Override
-    public Stream<RdbmsColumn> map(final TypeAttribute typeAttribute, RdbmsBuilder.RdbmsBuilderContext context) {
+    public Stream<RdbmsColumn> map(final TypeAttribute typeAttribute, RdbmsBuilderContext context) {
         final EMap<Node, EList<EClass>> ancestors = context.ancestors;
 
         final EClass sourceType = typeAttribute.getNode().getType();
-        sourceType.getEAllSuperTypes().forEach(superType -> {
+        for (EClass superType : sourceType.getEAllSuperTypes()) {
             log.trace("   - found super type: {}", AsmUtils.getClassifierFQName(superType));
             if (!ancestors.containsKey(typeAttribute.getNode())) {
                 ancestors.put(typeAttribute.getNode(), new UniqueEList<>());
             }
             // add ancestor for a given attribute
             ancestors.get(typeAttribute.getNode()).add(superType);
-        });
+        }
 
         return getTargets(typeAttribute)
                 .flatMap(t -> Arrays.asList(
