+++ judo-runtime-core-dao-rdbms/src/main/java/hu/blackbelt/judo/runtime/core/dao/rdbms/query/mappers/VariableMapper.java
@@ -23,13 +23,10 @@ package hu.blackbelt.judo.runtime.core.dao.rdbms.query.mappers;
 import hu.blackbelt.judo.meta.asm.runtime.AsmUtils;
 import hu.blackbelt.judo.meta.query.*;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.RdbmsBuilder;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.query.RdbmsBuilderContext;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.RdbmsField;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.RdbmsNamedParameter;
-import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import org.eclipse.emf.common.util.EList;
-import org.eclipse.emf.common.util.EMap;
-import org.eclipse.emf.ecore.EClass;
 import org.eclipse.emf.ecore.EEnum;
 import org.eclipse.emf.ecore.util.EcoreUtil;
 
@@ -44,7 +41,7 @@ public class VariableMapper<ID> extends RdbmsMapper<Variable> {
     public static final String PARAMETER_VARIABLE_KEY = "PARAMETER";
 
     @Override
-    public Stream<? extends RdbmsField> map(final Variable variable, RdbmsBuilder.RdbmsBuilderContext context) {
+    public Stream<? extends RdbmsField> map(final Variable variable, RdbmsBuilderContext context) {
         final RdbmsBuilder<?> rdbmsBuilder = context.rdbmsBuilder;
 
         final String id = EcoreUtil.getIdentification(variable);
@@ -63,7 +60,7 @@ public class VariableMapper<ID> extends RdbmsMapper<Variable> {
         }
     }
 
-    private Stream<? extends RdbmsField> getFields(final RdbmsBuilder.RdbmsBuilderContext context, final Variable variable) {
+    private Stream<? extends RdbmsField> getFields(final RdbmsBuilderContext context, final Variable variable) {
         final RdbmsBuilder<?> rdbmsBuilder = context.rdbmsBuilder;
         final Map<String, Object> queryParameters = context.queryParameters;
