+++ judo-runtime-core-dao-rdbms/src/main/java/hu/blackbelt/judo/runtime/core/dao/rdbms/query/processor/FilterJoinProcessor.java
@@ -1,21 +1,26 @@
-package hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.join;
+package hu.blackbelt.judo.runtime.core.dao.rdbms.query.processor;
 
 import hu.blackbelt.judo.meta.asm.runtime.AsmUtils;
 import hu.blackbelt.judo.meta.query.*;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.RdbmsResolver;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.executors.StatementExecutor;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.RdbmsBuilder;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.query.RdbmsBuilderContext;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.RdbmsField;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.RdbmsResultSet;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.join.RdbmsJoin;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.join.RdbmsQueryJoin;
+import hu.blackbelt.judo.runtime.core.dao.rdbms.query.model.join.RdbmsTableJoin;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.query.utils.RdbmsAliasUtil;
 import lombok.Builder;
 import lombok.NonNull;
+import lombok.extern.slf4j.Slf4j;
 import org.eclipse.emf.common.util.ECollections;
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.common.util.EMap;
 import org.eclipse.emf.common.util.UniqueEList;
 import org.eclipse.emf.ecore.EClass;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -23,30 +28,14 @@ import java.util.stream.Collectors;
 
 import static hu.blackbelt.judo.runtime.core.dao.rdbms.query.utils.RdbmsAliasUtil.getParentIdColumnAlias;
 
-public class AddFilterJoinAndConditions {
+@Builder
+@Slf4j
+public class FilterJoinProcessor<ID> {
+    @NonNull
+    private final RdbmsResolver rdbmsResolver;
 
-    @Builder
-    public static class AddFilterJoinsAndConditionsParameters {
-        @NonNull
-        List<RdbmsJoin> joins;
-        @NonNull
-        SubSelect query;
-        @NonNull
-        EList<Join> processedNodesForJoins;
-        @NonNull
-        List<RdbmsField> conditions;
-        @Builder.Default
-        String partnerTablePrefix = "";
-        @NonNull
-        Filter filter;
-        @NonNull
-        Node partnerTable;
-        boolean addJoinsOfFilterFeature;
-        @NonNull
-        RdbmsBuilder.RdbmsBuilderContext builderContext;
-    }
-    public static <ID> void addFilterJoinsAndConditions(AddFilterJoinsAndConditionsParameters params) {
-        final RdbmsBuilder<ID> rdbmsBuilder = (RdbmsBuilder<ID>) params.builderContext.rdbmsBuilder;
+    public void process(final FilterJoinProcessorParameters params, final RdbmsBuilderContext builderContext) {
+        final RdbmsBuilder<ID> rdbmsBuilder = (RdbmsBuilder<ID>) builderContext.rdbmsBuilder;
         final List<RdbmsJoin> joins = params.joins;
         final Filter filter = params.filter;
         final SubSelect query = params.query;
@@ -55,7 +44,6 @@ public class AddFilterJoinAndConditions {
         final String partnerTablePrefix = params.partnerTablePrefix;
         final Node partnerTable = params.partnerTable;
         final Boolean addJoinsOfFilterFeature = params.addJoinsOfFilterFeature;
-        final RdbmsBuilder.RdbmsBuilderContext builderContext = params.builderContext;
         final EMap<Node, EList<EClass>> ancestors = builderContext.ancestors;
         final EMap<Node, EList<EClass>> descendants = builderContext.descendants;
         final SubSelect parentIdFilterQuery = builderContext.parentIdFilterQuery;
@@ -70,6 +58,7 @@ public class AddFilterJoinAndConditions {
                     .partnerColumnName(partnerTable instanceof SubSelect ? getParentIdColumnAlias(query.getContainer()) : StatementExecutor.ID_COLUMN_NAME)
                     .alias(filter.getAlias())
                     .build());
+
             if (addJoinsOfFilterFeature) {
                 final EList<Join> navigationJoins = new UniqueEList<>();
                 final EList<Join> targetJoins = new UniqueEList<>(filter.getFeature().getNodes().stream()
@@ -88,33 +77,38 @@ public class AddFilterJoinAndConditions {
                 }
 
                 ECollections.reverse(navigationJoins);
-                navigationJoins.stream()
-                        .filter(join -> !processedNodesForJoins.contains(join))
-                        .forEach(join -> {
-                            joins.addAll(rdbmsBuilder.processJoin(
-                                    RdbmsBuilder.ProcessJoinParameters.builder()
-                                            .builderContext(builderContext)
-                                            .join(join)
-                                            .withoutFeatures(false)
-                                            .mask(null)
-                                            .build()
-                            ));
-                            processedNodesForJoins.add(join);
-                        });
-                filter.getFeature().getNodes().stream()
+
+                List<Join> unprocessedNodes = navigationJoins.stream()
+                        .filter(join -> !processedNodesForJoins.contains(join)).collect(Collectors.toList());
+
+                for (Join join : unprocessedNodes) {
+                    joins.addAll(rdbmsBuilder.processJoin(
+                            JoinProcessParameters.builder()
+                                    .builderContext(builderContext)
+                                    .join(join)
+                                    .withoutFeatures(false)
+                                    .mask(null)
+                                    .build()
+                    ));
+                    processedNodesForJoins.add(join);
+                }
+
+                List<Join> allMissedJoins = filter.getFeature().getNodes().stream()
                         .filter(n -> !processedNodesForJoins.contains(n) && !AsmUtils.equals(n, filter) && n instanceof Join)
                         .flatMap(n -> ((Join) n).getAllJoins().stream())
-                        .forEach(join -> {
-                            joins.addAll(rdbmsBuilder.processJoin(
-                                    RdbmsBuilder.ProcessJoinParameters.builder()
-                                            .builderContext(builderContext)
-                                            .join(join)
-                                            .withoutFeatures(false)
-                                            .mask(null)
-                                            .build()
-                            ));
-                            processedNodesForJoins.add(join);
-                        });
+                        .collect(Collectors.toList());
+
+                for (Join join : allMissedJoins) {
+                    joins.addAll(rdbmsBuilder.processJoin(
+                            JoinProcessParameters.builder()
+                                    .builderContext(builderContext)
+                                    .join(join)
+                                    .withoutFeatures(false)
+                                    .mask(null)
+                                    .build()
+                    ));
+                    processedNodesForJoins.add(join);
+                }
             }
         }
